SXUSecondPlatform-Final/requirements.txt
alembic==0.9.9
click==6.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysql-connector==2.1.6
numpy==1.14.3
Pillow==5.1.0
PyMySQL==0.8.1
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
WTForms==2.2
